Fix problems seen during session : .gitignore / .idea, namespace for router, fontAwesome webfonts 404 , Folder structure reorg, move views in App folder
