+++ .gitignore
@@ -1,5 +1,6 @@
 composer.phar
 /vendor/
+/.idea/
 
 # Commit your application's lock file https://getcomposer.org/doc/01-basic-usage.md#commit-your-composer-lock-file-to-version-control
 # You may choose to ignore a library lock file http://getcomposer.org/doc/02-libraries.md#lock-file
