--- App/Database.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace App;
-
-class Database
-{
-    public function __construct()
-    {
-        echo "App\Database";
-    }
-}
