+++ App/Database/Database.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Database;
+
+class Connection
+{
+    public function __construct()
+    {
+        echo "App\Database\Connection";
+    }
+}
