+++ public/index.php
@@ -1,7 +1,6 @@
 <?php
-
-namespace App;
 require __DIR__ . '/../vendor/autoload.php';
+use App\Routing\Router;
 $route = new Router();
 $route->run();
 ?>
